# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: 狠狠学 HTTP API
  version: 1.0.0
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: |
    第一期开工前的规范性 HTTP 契约。所有响应使用 ApiResponse 包装；
    字段约束、业务状态机和补充示例见 03-api-specification.md。
servers:
  - url: /api/v1
security:
  - SaToken: []
paths:
  /auth/register:
    post:
      security: []
      summary: Register
      operationId: register
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {$ref: '#/components/requestBodies/RegisterRequest'}
      responses: {'201': {$ref: '#/components/responses/User'}, '400': {$ref: '#/components/responses/Error'}, '409': {$ref: '#/components/responses/Error'}}
  /auth/login:
    post:
      security: []
      summary: Login
      operationId: login
      requestBody: {$ref: '#/components/requestBodies/LoginRequest'}
      responses: {'200': {$ref: '#/components/responses/Login'}, '400': {$ref: '#/components/responses/Error'}, '401': {$ref: '#/components/responses/Error'}, '429': {$ref: '#/components/responses/Error'}, '503': {$ref: '#/components/responses/Error'}}
  /auth/email-codes:
    post:
      security: []
      summary: Send email code
      operationId: sendEmailCode
      requestBody: {$ref: '#/components/requestBodies/EmailCodeRequest'}
      responses: {'202': {$ref: '#/components/responses/Empty'}, '400': {$ref: '#/components/responses/Error'}, '409': {$ref: '#/components/responses/Error'}, '429': {$ref: '#/components/responses/Error'}}
...
  /auth/logout:
    post:
      summary: Logout
      operationId: logout
      responses: {'204': {$ref: '#/components/responses/NoContent'}, '400': {$ref: '#/components/responses/Error'}}
...
components:
  securitySchemes:
    SaToken: {type: apiKey, in: header, name: satoken}
  parameters:
    IdempotencyKey: {name: Idempotency-Key, in: header, required: true, schema: {type: string, format: uuid}}
...
  requestBodies:
    RegisterRequest: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/RegisterRequest'}}}}
    LoginRequest: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/LoginRequest'}}}}
    EmailCodeRequest: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/EmailCodeRequest'}}}}
...
  responses:
    NoContent: {description: Completed without a response body}
    Empty: {$ref: '#/components/responses/ApiResponse'}
    Error: {description: Error response, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
    ApiResponse: {description: Success response, content: {application/json: {schema: {$ref: '#/components/schemas/ApiResponse'}}}}
    Login: {description: Login session, content: {application/json: {schema: {$ref: '#/components/schemas/LoginResponse'}}}}
    User: {description: User, headers: {ETag: {$ref: '#/components/headers/ETag'}}, content: {application/json: {schema: {$ref: '#/components/schemas/UserResponse'}}}}
...
  headers:
    ETag: {schema: {type: string, pattern: '^"[1-9][0-9]*"$'}, description: New aggregate version}
  schemas:
    Ulid: {type: string, pattern: '^(?:[0-9]{26}|[0-9A-HJKMNP-TV-Z]{26})$', description: '新 ID 为 26 位十进制雪花 ID；兼容已存在的 26 位 ULID。'}
...
    ApiResponse: {type: object, required: [code, message, data, traceId], properties: {code: {const: 0}, message: {const: OK}, data: {}, traceId: {type: string}}}
    ErrorResponse: {type: object, required: [code, message, traceId], properties: {code: {type: string, description: '业务错误码；INTERNAL_ERROR 表示未预期服务端错误，SERVICE_UNAVAILABLE 表示服务暂不可用。'}, message: {type: string}, data: {type: 'null'}, traceId: {type: string}}}
    RegisterRequest: {type: object, additionalProperties: false, required: [username, email, emailCode, password, nickname], properties: {username: {type: string, minLength: 8, maxLength: 64}, email: {type: string, format: email}, emailCode: {type: string, pattern: '^\\d{6}$'}, password: {type: string, minLength: 8, maxLength: 128, pattern: '^(?=.*[A-Za-z])(?=.*\\d).+$'}, nickname: {type: string, minLength: 1, maxLength: 64}}}
    LoginRequest: {type: object, additionalProperties: false, required: [account, password], properties: {account: {type: string, minLength: 1, maxLength: 128}, password: {type: string, minLength: 8, maxLength: 128}}}
    EmailCodeRequest: {type: object, additionalProperties: false, required: [email, purpose], properties: {email: {type: string, format: email}, purpose: {type: string, enum: [REGISTER, PASSWORD_RESET]}}}
...
    LoginResponse: {allOf: [{$ref: '#/components/schemas/ApiResponse'}, {properties: {data: {$ref: '#/components/schemas/Login'}}}]}
    UserResponse: {allOf: [{$ref: '#/components/schemas/ApiResponse'}, {properties: {data: {$ref: '#/components/schemas/User'}}}]}
...
    Login: {type: object, required: [token, user], properties: {token: {type: string}, user: {$ref: '#/components/schemas/User'}}}
    User: {type: object, required: [id, username, email, nickname, permissions, version], properties: {id: {$ref: '#/components/schemas/Ulid'}, username: {type: string}, email: {type: string, format: email}, emailVerifiedAt: {type: [string, 'null'], format: date-time}, nickname: {type: string}, avatarFileId: {oneOf: [{$ref: '#/components/schemas/Ulid'}, {type: 'null'}]}, permissions: {type: array, items: {type: string}}, version: {type: integer, minimum: 1}}}
...