# backend/services/auth-service/src/main/java/com/hengxue/auth/interfaces/rest/request/LoginRequest.java
package com.hengxue.auth.interfaces.rest.request;

import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Pattern;
import jakarta.validation.constraints.Size;

/** 密码登录的 HTTP 请求体。 */
@JsonIgnoreProperties(ignoreUnknown = false)
public record LoginRequest(
        @NotBlank @Size(max = 128) String account,
        @NotBlank @Size(min = 8, max = 128)
        @Pattern(regexp = "^(?=.*[A-Za-z])(?=.*\\d).+$") String password
) {
}